Collect name clashes into haiku-include-before.h and use it.
It is to be included before Haiku headers and supposed to set up
things so that no name clashes between Haiku and QEMU APIs happen.
Right now it redefines load_image as QEMU has its own version so we
don't want the Haiku version to collide, and it undefs the atomic
functions that would otherwise trip up the ones in Haiku.